# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Chocadies Recruiting API
  version: "2.0.0"
  description: |
    Interne Recruiting-API der Chocolatier Chocadies AG.
    - Auth via **JWT Bearer** (Admin/HR/Applicant)
    - Partner-API via **x-api-key** (Agencies)
servers:
  - url: /
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agencyKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    MeetingPublic:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [proposed, accepted, cancelled] }
        startsAt: { type: string, format: date-time }
        endsAt:   { type: string, format: date-time }
        mode: { type: string, enum: [online, onsite] }
        location: { type: string }
        job:
          type: object
          properties:
            id: { type: string }
            title: { type: string }
    MeetingAdmin:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [proposed, accepted, cancelled] }
        startsAt: { type: string, format: date-time }
        endsAt:   { type: string, format: date-time }
        mode: { type: string, enum: [online, onsite] }
        location: { type: string }
        applicant:
          type: object
          properties:
            email: { type: string, format: email }
...
paths:
...
  /api/v2/meetings:
    post:
      tags: [Meetings]
      security: [{ bearerAuth: [] }]
      summary: Meeting vorschlagen/erstellen
      description: |
        Applicant: { jobId, startsAt, endsAt, mode, location }
        Admin/HR kann zusätzlich applicantEmail angeben.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                jobId: { type: string }
                applicantEmail: { type: string }
                startsAt: { type: string, format: date-time }
                endsAt:   { type: string, format: date-time }
                mode: { type: string, enum: [online, onsite] }
                location: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  id: { type: string }
                  status: { type: string }

    get:
      tags: [Meetings]
      security: [{ bearerAuth: [] }]
      summary: Meetings zu Job (Admin/HR)
      parameters:
        - in: query
          name: jobId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Liste Admin-View
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MeetingAdmin' }
  /api/v2/meetings/me:
    get:
      tags: [Meetings]
      security: [{ bearerAuth: [] }]
      summary: Meine Meetings (Applicant)
      responses:
        '200':
          description: Liste Applicant-View
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MeetingPublic' }
  /api/v2/meetings/{id}:
    patch:
      tags: [Meetings]
      security: [{ bearerAuth: [] }]
      summary: Meeting aktualisieren (Admin/HR)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [proposed, accepted, cancelled] }
                location: { type: string }
      responses:
        '200': { description: ok }
    delete:
      tags: [Meetings]
      security: [{ bearerAuth: [] }]
      summary: Meeting löschen (Admin/HR)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: ok }
...
  /api/v2/linkedin/publish/{jobId}:
    delete:
      tags: [LinkedIn (dummy)]
      security: [{ bearerAuth: [] }]
      summary: Dummy-Unpublish
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        '200': { description: ok }
  /api/v2/linkedin/import-applicant:
    post:
      tags: [LinkedIn (dummy)]
      security: [{ bearerAuth: [] }]
      summary: Dummy-Import eines Bewerbers (optional inkl. Bewerbung)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profileUrl, email, name]
              properties:
                profileUrl: { type: string }
                email: { type: string, format: email }
                name: { type: string }
                birthdate: { type: string, format: date }
                address: { type: string }
                jobId: { type: string, description: "Optional: legt sofort Bewerbung an" }
      responses:
        '200':
          description: ok + IDs
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  userId: { type: string }
                  applicationId: { type: string, nullable: true }
...